+++ apps/mobile/src/components/posts/footer/index.tsx
@@ -14,10 +14,11 @@ import { PostMeta } from './meta'
 type Props = {
   community?: boolean
   post: Post
+  privacy?: boolean
   style?: StyleProp<ViewStyle>
 }
 
-export function PostFooter({ community = true, post, style }: Props) {
+export function PostFooter({ community = true, post, privacy, style }: Props) {
   const a11y = useTranslations('a11y')
 
   const { hidePostActions } = usePreferences(['hidePostActions'])
@@ -29,14 +30,14 @@ export function PostFooter({ community = true, post, style }: Props) {
       <View style={styles.header}>
         {community ? <PostCommunity post={post} /> : null}
 
-        <PostMeta post={post} />
+        <PostMeta post={post} privacy={privacy} />
       </View>
 
       {hidePostActions ? null : (
         <View style={styles.footer}>
           <FooterButton
-            color={post.liked === true ? 'orange' : undefined}
-            fill={post.liked === true}
+            color={!privacy && post.liked === true ? 'orange' : undefined}
+            fill={!privacy && post.liked === true}
             icon={getIcon('upvote.fill')}
             label={a11y(post.liked ? 'removeUpvote' : 'upvote')}
             onPress={() => {
@@ -48,8 +49,8 @@ export function PostFooter({ community = true, post, style }: Props) {
           />
 
           <FooterButton
-            color={post.liked === false ? 'violet' : undefined}
-            fill={post.liked === false}
+            color={!privacy && post.liked === false ? 'violet' : undefined}
+            fill={!privacy && post.liked === false}
             icon={getIcon('downvote.fill')}
             label={a11y(post.liked === false ? 'removeDownvote' : 'downvote')}
             onPress={() => {